gpu: nvgpu: common/pmu: fix compile error of new compile flags
It's preparing to add bellow CFLAGS:
    -Werror -Wall -Wextra \
    -Wmissing-braces -Wpointer-arith -Wundef \
    -Wconversion -Wsign-conversion \
    -Wformat-security \
    -Wmissing-declarations -Wredundant-decls -Wimplicit-fallthrough

Jira GVSCI-11640

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvgpu/+/2555055
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ drivers/gpu/nvgpu/common/pmu/pmu_rtos_init.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2021, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2017-2022, NVIDIA CORPORATION.  All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -249,7 +249,7 @@ void nvgpu_pmu_rtos_cmdline_args_init(struct gk20a *g, struct nvgpu_pmu *pmu)
 		pmu, GK20A_PMU_DMAIDX_VIRT);
 
 	pmu->fw->ops.set_cmd_line_args_cpu_freq(pmu,
-		g->ops.clk.get_rate(g, CTRL_CLK_DOMAIN_PWRCLK));
+		(u32)g->ops.clk.get_rate(g, CTRL_CLK_DOMAIN_PWRCLK));
 
 	if (pmu->fw->ops.config_cmd_line_args_super_surface != NULL) {
 		pmu->fw->ops.config_cmd_line_args_super_surface(pmu);
